fixed bug in ajaxPurgeConfigs.php that caused it to fail

+++ www/lib/ajaxHandlers/ajaxPurgeConfigs.php
@@ -1,6 +1,7 @@
 <?php
 
 require_once("/home/rconfig/classes/usersession.class.php");
+require_once("/home/rconfig/config/functions.inc.php");
 require_once("/home/rconfig/classes/ADLog.class.php");
 $log = ADLog::getInstance();
 if (!$session->logged_in) {
@@ -25,9 +26,10 @@ if (!$session->logged_in) {
 // tasks: delete all files entries in DB by ID and then delete all Dirs
 // 1. get all ids from DB older than X days
     $db2 = new db2();
-    $db2->query("SELECT id FROM configs WHERE DATE_SUB(CURDATE(),INTERVAL :purgeDays DAY) >= configDate");
-    $db2->bind(':purgeDays', $purgeDays); //bind here and create wildcard search term here also
-    $getIDs = $db2->resultsetCols();
+    $db2->query("SELECT id FROM configs WHERE DATE_SUB(CURDATE(),INTERVAL " . $purgeDays . " DAY) >= configDate");
+    $getIDs = $db2->resultset();
+    $getIDs = flatten($getIDs);
+        
 // logging below
     if ($getIDs) {
         $log->Info("Info: Start manual " . $purgeDays . " day Config File Purge - GET DB IDs(File: " . $_SERVER['PHP_SELF'] . ")");
@@ -38,39 +40,40 @@ if (!$session->logged_in) {
     $iDlist = implode(", ", $getIDs);
 
 // 2. get all dirs using a group by from DB older than X days
-//$getDirs = 'SELECT configLocation FROM configs WHERE DATE_SUB(CURDATE(),INTERVAL '.$purgeDays.' DAY) >= configDate GROUP BY configLocation;';
-    $db2->query("SELECT configLocation FROM configs WHERE DATE_SUB(CURDATE(),INTERVAL :purgeDays DAY) >= configDate GROUP BY configLocation;");
-    $db2->bind(':purgeDays', $purgeDays); //bind here and create wildcard search term here also
+    $db2->query("SELECT configLocation FROM configs WHERE DATE_SUB(CURDATE(),INTERVAL " . $purgeDays . " DAY) >= configDate GROUP BY configLocation;");
     $getDirRes = $db2->resultsetCols();
-    if ($executeRes) {
-        $log->Info("Info: Start manual " . $purgeDays . " day Config File Purge - GET DIRs (File: " . $_SERVER['PHP_SELF'] . ")");
-    } else {
-        $response['response'] = "ERROR: executing query " . $getDirs;
-        $log->Fatal("Fatal: Could not Config File Purge - GET DIRs (File: " . $_SERVER['PHP_SELF'] . ")");
-    }
-// physically remove directories
-    $dirsToPurgeArr = array();
-    while ($row = $getDirRes) {
-        foreach ($row as $k => $v) {
-            exec('rm -fr ' . $v);
-        }
-    }
-// delete all empty Dirs under /home/rconfig/data/ for completeness
-    exec('find /home/rconfig/data/. -type d -empty -delete');
 
+    if (count($getDirRes) > 0) {
+        $log->Info("Info: Start manual " . $purgeDays . " day Config File Purge - GET DIRs (File: " . $_SERVER['PHP_SELF'] . ")");
+
+// physically remove directories
+        foreach ($getDirRes as $row) {
+            foreach ($row as $k => $v) {
+                echo $v;
+                exec('rm -fr ' . $v);
+            }
+        }
+// delete all empty Dirs under /home/rconfig/data/ for completeness
+        exec('find /home/rconfig/data/. -type d -empty -delete');
+        
 // 3. Delete all ids from DB older than X days - if these are delete first - step 2 cannot 
 //    work as will not be able to get unique dirs older than X days
-    if (!empty($iDlist)) {
-        $db2->query("DELETE FROM configs WHERE id IN ($iDlist)");
-        $executeRes = $db2->execute();
-        if ($executeRes) {
-            $log->Info("Info: Start manual " . $purgeDays . " day Config File Purge - Delete DB Rows (File: " . $_SERVER['PHP_SELF'] . ")");
-            $response['response'] = "SUCCESS: " . $purgeDays . " day Config File Purge Completed";
-        } else {
-            $response['response'] = "ERROR: executing query Delete ID";
-            $log->Fatal("Fatal: Could not Config File Purge - Delete DB Rows (File: " . $_SERVER['PHP_SELF'] . ")");
-        }
-    }
 
-    echo json_encode($response);
+        if (!empty($iDlist)) {
+            $db2->query("DELETE FROM configs WHERE id IN ($iDlist)");
+            $executeRes = $db2->execute();
+            if ($executeRes) {
+                $log->Info("Info: Start manual " . $purgeDays . " day Config File Purge - Delete DB Rows (File: " . $_SERVER['PHP_SELF'] . ")");
+                $response['response'] = "SUCCESS: " . $purgeDays . " day Config File Purge Completed";
+            } else {
+                $response['response'] = "ERROR: executing query Delete ID";
+                $log->Fatal("Fatal: Could not Config File Purge - Delete DB Rows (File: " . $_SERVER['PHP_SELF'] . ")");
+            }
+        }
+
+        echo json_encode($response);
+    } else {
+        $response['response'] = "ERROR: executing query ";
+        $log->Fatal("Fatal: Could not Config File Purge - GET DIRs (File: " . $_SERVER['PHP_SELF'] . ")");
+    }
 }
